@@ -1,8 +1,9 @@
 import asyncio
 import json
+from collections.abc import AsyncGenerator
 from sqlite3 import PARSE_DECLTYPES, ProgrammingError
 from types import TracebackType
-from typing import Any, AsyncGenerator, Dict, List, Optional, Set
+from typing import Any
 from urllib.parse import urlsplit
 from uuid import uuid4
 
@@ -40,7 +41,7 @@ class Transaction(TransactionABC):
     def __init__(self, connection: "Connection", *, force_rollback: bool = False) -> None:
         self._connection = connection
         self._force_rollback = force_rollback
-        self._savepoints: List[str] = []
+        self._savepoints: list[str] = []
 
     async def __aenter__(self) -> "Transaction":
         await self.start()
@@ -85,14 +86,14 @@ class Connection(ConnectionABC):
         self._connection = connection
         self._lock = asyncio.Lock()
 
-    async def execute(self, query: LiteralString, values: Optional[ValueType] = None) -> None:
+    async def execute(self, query: LiteralString, values: ValueType | None = None) -> None:
         try:
             async with self._lock:
                 await self._connection.execute(query, values)
         except ProgrammingError as error:
             raise UndefinedParameterError(str(error))
 
-    async def execute_many(self, query: LiteralString, values: List[ValueType]) -> None:
+    async def execute_many(self, query: LiteralString, values: list[ValueType]) -> None:
         if not values:
             return
 
@@ -105,8 +106,8 @@ class Connection(ConnectionABC):
     async def fetch_all(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> List[RecordType]:
+        values: ValueType | None = None,
+    ) -> list[RecordType]:
         try:
             async with self._lock:
                 async with self._connection.execute(query, values) as cursor:
@@ -119,8 +120,8 @@ class Connection(ConnectionABC):
     async def fetch_first(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         try:
             async with self._lock:
                 async with self._connection.execute(query, values) as cursor:
@@ -136,15 +137,15 @@ class Connection(ConnectionABC):
     async def fetch_one(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         return await self.fetch_first(query, values)
 
     async def fetch_sole(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         try:
             async with self._lock:
                 async with self._connection.execute(query, values) as cursor:
@@ -162,8 +163,8 @@ class Connection(ConnectionABC):
     async def fetch_val(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[Any]:
+        values: ValueType | None = None,
+    ) -> Any | None:
         try:
             async with self._lock:
                 async with self._connection.execute(query, values) as cursor:
@@ -178,7 +179,7 @@ class Connection(ConnectionABC):
     async def iterate(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
+        values: ValueType | None = None,
     ) -> AsyncGenerator[RecordType, None]:
         try:
             async with self._lock:
@@ -193,11 +194,11 @@ class Connection(ConnectionABC):
 
 
 class Backend(BackendABC):
-    def __init__(self, url: str, options: Dict[str, Any], type_converters: TypeConverters) -> None:
+    def __init__(self, url: str, options: dict[str, Any], type_converters: TypeConverters) -> None:
         _, _, path, *_ = urlsplit(url)
         self._path = path[1:]
         self._options = options
-        self._connections: Set[aiosqlite.Connection] = set()
+        self._connections: set[aiosqlite.Connection] = set()
         for _, converters in {**DEFAULT_TYPE_CONVERTERS, **type_converters}.items():
             for typename, (encoder, decoder, pytype) in converters.items():
                 aiosqlite.register_adapter(pytype, encoder)
@@ -206,7 +207,7 @@ class Backend(BackendABC):
     async def connect(self) -> None:
         pass
 
-    async def disconnect(self, timeout: Optional[int] = None) -> None:
+    async def disconnect(self, timeout: int | None = None) -> None:
         tasks = [asyncio.ensure_future(connection.close()) for connection in self._connections]
         await asyncio.wait_for(asyncio.gather(*tasks), timeout)
 
@@ -240,7 +241,7 @@ class Backend(BackendABC):
 
 
 class TestingBackend(BackendABC):
-    def __init__(self, url: str, options: Dict[str, Any], type_converters: TypeConverters) -> None:
+    def __init__(self, url: str, options: dict[str, Any], type_converters: TypeConverters) -> None:
         _, _, path, *_ = urlsplit(url)
         self._path = path[1:]
         self._options = options
@@ -259,7 +260,7 @@ class TestingBackend(BackendABC):
         await connection.__aenter__()
         self._connection = Connection(connection)
 
-    async def disconnect(self, timeout: Optional[int] = None) -> None:
+    async def disconnect(self, timeout: int | None = None) -> None:
         await asyncio.wait_for(self._connection._connection.close(), timeout)
 
     async def acquire(self) -> Connection:
