@@ -1,7 +1,8 @@
 import asyncio
 import json
+from collections.abc import AsyncGenerator
 from types import TracebackType
-from typing import Any, AsyncGenerator, Dict, List, Optional, Tuple
+from typing import Any
 
 import asyncpg
 from buildpg import BuildError, render
@@ -38,7 +39,7 @@ DEFAULT_TYPE_CONVERTERS: TypeConverters = {
 class Transaction(TransactionABC):
     def __init__(self, connection: "Connection", *, force_rollback: bool = False) -> None:
         self._connection = connection
-        self._transaction: Optional[asyncpg.Transaction] = None
+        self._transaction: asyncpg.Transaction | None = None
         self._force_rollback = force_rollback
 
     async def __aenter__(self) -> "Transaction":
@@ -74,7 +75,7 @@ class Connection(ConnectionABC):
         self._connection = connection
         self._lock = asyncio.Lock()
 
-    async def execute(self, query: LiteralString, values: Optional[ValueType] = None) -> None:
+    async def execute(self, query: LiteralString, values: ValueType | None = None) -> None:
         compiled_query, args = self._compile(query, values)
         try:
             async with self._lock:
@@ -82,7 +83,7 @@ class Connection(ConnectionABC):
         except asyncpg.exceptions.UndefinedParameterError as error:
             raise UndefinedParameterError(str(error))
 
-    async def execute_many(self, query: LiteralString, values: List[ValueType]) -> None:
+    async def execute_many(self, query: LiteralString, values: list[ValueType]) -> None:
         if not values:
             return
 
@@ -98,8 +99,8 @@ class Connection(ConnectionABC):
     async def fetch_all(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> List[RecordType]:
+        values: ValueType | None = None,
+    ) -> list[RecordType]:
         compiled_query, args = self._compile(query, values)
         try:
             async with self._lock:
@@ -110,8 +111,8 @@ class Connection(ConnectionABC):
     async def fetch_first(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         compiled_query, args = self._compile(query, values)
         try:
             async with self._lock:
@@ -123,15 +124,15 @@ class Connection(ConnectionABC):
     async def fetch_one(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         return await self.fetch_first(query, values)
 
     async def fetch_sole(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
-    ) -> Optional[RecordType]:
+        values: ValueType | None = None,
+    ) -> RecordType | None:
         compiled_query, args = self._compile(query, values)
         try:
             async with self._lock:
@@ -149,7 +150,7 @@ class Connection(ConnectionABC):
     async def fetch_val(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
+        values: ValueType | None = None,
     ) -> Any:
         compiled_query, args = self._compile(query, values)
         try:
@@ -161,7 +162,7 @@ class Connection(ConnectionABC):
     async def iterate(
         self,
         query: LiteralString,
-        values: Optional[ValueType] = None,
+        values: ValueType | None = None,
     ) -> AsyncGenerator[RecordType, None]:
         compiled_query, args = self._compile(query, values)
         async with self._lock:
@@ -176,8 +177,8 @@ class Connection(ConnectionABC):
         return Transaction(self, force_rollback=force_rollback)
 
     def _compile(
-        self, query: LiteralString, values: Optional[ValueType] = None
-    ) -> Tuple[str, List[Any]]:
+        self, query: LiteralString, values: ValueType | None = None
+    ) -> tuple[str, list[Any]]:
         if isinstance(values, dict):
             try:
                 return render(query, **(values or {}))
@@ -190,8 +191,8 @@ class Connection(ConnectionABC):
 
 
 class Backend(BackendABC):
-    def __init__(self, url: str, options: Dict[str, Any], type_converters: TypeConverters) -> None:
-        self._pool: Optional[asyncpg.Pool] = None
+    def __init__(self, url: str, options: dict[str, Any], type_converters: TypeConverters) -> None:
+        self._pool: asyncpg.Pool | None = None
         self._url = url
         self._options = options
         self._type_converters = {**DEFAULT_TYPE_CONVERTERS, **type_converters}
@@ -200,7 +201,7 @@ class Backend(BackendABC):
         if self._pool is None:
             self._pool = await asyncpg.create_pool(dsn=self._url, init=self._init, **self._options)
 
-    async def disconnect(self, timeout: Optional[int] = None) -> None:
+    async def disconnect(self, timeout: int | None = None) -> None:
         if self._pool is not None:
             await asyncio.wait_for(self._pool.close(), timeout)
         self._pool = None
@@ -225,7 +226,7 @@ class Backend(BackendABC):
 
 
 class TestingBackend(BackendABC):
-    def __init__(self, url: str, options: Dict[str, Any], type_converters: TypeConverters) -> None:
+    def __init__(self, url: str, options: dict[str, Any], type_converters: TypeConverters) -> None:
         self._url = url
         self._options = options
         self._type_converters = {**DEFAULT_TYPE_CONVERTERS, **type_converters}
@@ -234,7 +235,7 @@ class TestingBackend(BackendABC):
         self._connection = Connection(await asyncpg.connect(dsn=self._url, **self._options))
         await _init_connection(self._connection._connection, self._type_converters)
 
-    async def disconnect(self, timeout: Optional[int] = None) -> None:
+    async def disconnect(self, timeout: int | None = None) -> None:
         await asyncio.wait_for(self._connection._connection.close(), timeout)
 
     async def acquire(self) -> Connection:
